Lots a shit. Started asg5.

+++ asg5/lyutils.h
@@ -18,7 +18,7 @@ extern int yyin_linenr;
 extern char* yytext;
 extern int yy_flex_debug;
 extern int yydebug;
-extern int yyleng;
+extern unsigned long yyleng;
 
 extern ofstream tok_file;
